Add files via upload
Pizza Sales AnalysisPizza Sales Analysis
I have created this project and analysis pizza sales data with SSMS and created a interactive dashboard with Excel.

firstly imported the relaed data into SSMS (SQL Server Management Studio) and created a database to store it. From there, i used SQL queries to analyze the data and extract insights such as the most popular pizza or the busiest times of day for sales.
after analysising the data through queries i created a dashboard with the help of Excel, i used the data i analyzed in SSMS to create charts and graphs that visualize the information in an easy-to-understand way. I also used Excel's built-in tools to create pivot tables and other interactive features that allows to explore the data in more detail.

Overall, creating a project on pizza sales analysis with SSMS and Excel can be a great way to gain insights into business and make data-driven decisions that can help in improving sales and profitabilit